+++ src/backend/azure.rs
@@ -1,6 +1,5 @@
 //! Implementation of the Azure Blob Storage backend.
 
-use std::ops::Range;
 use std::sync::Arc;
 
 use base64::Engine;
@@ -8,10 +7,10 @@ use base64::prelude::BASE64_STANDARD;
 use chrono::Utc;
 use crc64fast_nvme::Digest;
 use reqwest::Body;
-use reqwest::Client;
 use reqwest::Response;
 use reqwest::StatusCode;
 use reqwest::header;
+use reqwest_middleware::ClientWithMiddleware;
 use serde::Deserialize;
 use serde::Serialize;
 use tokio::sync::broadcast;
@@ -178,7 +177,7 @@ impl ResponseExt for Response {
 /// Represents an upload of a blob to Azure Blob Storage.
 pub struct AzureBlobUpload {
     /// The HTTP client to use for the upload.
-    client: Client,
+    client: ClientWithMiddleware,
     /// The blob URL.
     url: Url,
     /// The Azure block id.
@@ -190,7 +189,7 @@ pub struct AzureBlobUpload {
 impl AzureBlobUpload {
     /// Constructs a new blob upload.
     fn new(
-        client: Client,
+        client: ClientWithMiddleware,
         url: Url,
         block_id: Arc<String>,
         events: Option<broadcast::Sender<TransferEvent>>,
@@ -237,6 +236,7 @@ impl Upload for AzureBlobUpload {
             ByteStream::new(bytes),
             id,
             block,
+            0,
             self.events.clone(),
         ));
 
@@ -302,7 +302,7 @@ pub struct AzureBlobStorageBackend {
     /// The config to use for transferring files.
     config: Config,
     /// The HTTP client to use for transferring files.
-    client: Client,
+    client: ClientWithMiddleware,
     /// The channel for sending transfer events.
     events: Option<broadcast::Sender<TransferEvent>>,
 }
@@ -311,9 +311,10 @@ impl AzureBlobStorageBackend {
     /// Constructs a new Azure Blob Storage backend with the given configuration
     /// and events channel.
     pub fn new(config: Config, events: Option<broadcast::Sender<TransferEvent>>) -> Self {
+        let client = new_http_client(&config);
         Self {
             config,
-            client: new_http_client(),
+            client,
             events,
         }
     }
@@ -500,7 +501,7 @@ impl StorageBackend for AzureBlobStorageBackend {
         Ok(response)
     }
 
-    async fn get_range(&self, url: Url, etag: &str, range: Range<u64>) -> Result<Response> {
+    async fn get_at_offset(&self, url: Url, etag: &str, offset: u64) -> Result<Response> {
         debug_assert!(
             Self::is_supported_url(&self.config, &url),
             "{url} is not a supported Azure URL",
@@ -508,10 +509,8 @@ impl StorageBackend for AzureBlobStorageBackend {
         );
 
         debug!(
-            "sending ranged GET request for `{url}` ({start}-{end})",
+            "sending GET request at offset {offset} for `{url}`",
             url = url.display(),
-            start = range.start,
-            end = range.end
         );
 
         let response = self
@@ -519,10 +518,7 @@ impl StorageBackend for AzureBlobStorageBackend {
             .get(url)
             .header(header::USER_AGENT, USER_AGENT)
             .header(header::DATE, Utc::now().to_rfc2822())
-            .header(
-                header::RANGE,
-                format!("bytes={start}-{end}", start = range.start, end = range.end),
-            )
+            .header(header::RANGE, format!("bytes={offset}-"))
             .header(header::IF_MATCH, etag)
             .header(AZURE_VERSION_HEADER, AZURE_STORAGE_VERSION)
             .send()
@@ -535,14 +531,9 @@ impl StorageBackend for AzureBlobStorageBackend {
             return Err(Error::RemoteContentModified);
         }
 
-        // Handle error response
+        // Handle other error responses
         if !status.is_success() {
             return Err(response.into_error().await);
-        }
-
-        // We expect partial content, otherwise treat as remote content modified
-        if status != StatusCode::PARTIAL_CONTENT {
-            return Err(Error::RemoteContentModified);
         }
 
         Ok(response)