+++ src/transfer.rs
@@ -1,9 +1,8 @@
 //! Implementation of file transfers.
 
 use std::fs::File;
-use std::ops::Range;
+use std::io::SeekFrom;
 use std::path::Path;
-use std::path::PathBuf;
 use std::sync::Arc;
 use std::sync::atomic::AtomicU64;
 use std::sync::atomic::Ordering;
@@ -11,18 +10,17 @@ use std::sync::atomic::Ordering;
 use bytes::Bytes;
 use futures::FutureExt;
 use futures::StreamExt;
-use futures::TryFutureExt;
 use futures::TryStreamExt;
 use futures::stream;
+use reqwest::StatusCode;
 use reqwest::header;
 use tempfile::NamedTempFile;
 use tokio::fs;
-use tokio::io::AsyncReadExt;
+use tokio::io::AsyncSeekExt;
+use tokio::io::AsyncWriteExt;
 use tokio::io::BufWriter;
 use tokio::select;
-use tokio::task::spawn_blocking;
 use tokio_retry2::Retry;
-use tokio_retry2::RetryError;
 use tokio_util::io::StreamReader;
 use tokio_util::sync::CancellationToken;
 use tracing::debug;
@@ -37,40 +35,8 @@ use crate::UrlExt;
 use crate::backend::StorageBackend;
 use crate::backend::Upload;
 use crate::notify_retry;
-use crate::pool::BufferGuard;
 use crate::pool::BufferPool;
 use crate::streams::TransferStream;
-
-/// Represents information about a file being downloaded.
-#[derive(Clone)]
-struct DownloadInfo {
-    /// The file transfer id.
-    id: u64,
-    /// The size of the file being downloaded.
-    file_size: u64,
-    /// The block size to use for the download.
-    block_size: u64,
-    /// The number of blocks in the file.
-    num_blocks: u64,
-    /// The ETAG of the file being downloaded.
-    etag: Arc<String>,
-}
-
-/// Represents information about a block being downloaded.
-struct BlockDownloadInfo {
-    /// The file transfer id.
-    id: u64,
-    /// The ETAG of the resource being downloaded.
-    etag: Arc<String>,
-    /// The file being written.
-    file: Arc<File>,
-    /// The block number being downloaded.
-    block: u64,
-    /// The buffer to use to read and write the block.
-    buffer: BufferGuard,
-    /// The range in the file being downloaded.
-    range: Range<u64>,
-}
 
 /// Represents information about a file being uploaded.
 #[derive(Clone, Copy)]
@@ -104,15 +70,55 @@ where
         self.next_id.fetch_add(1, Ordering::SeqCst)
     }
 
-    /// Downloads a file using a single block.
+    /// Downloads a file to the given destination path.
     async fn download(
         &self,
-        id: u64,
         source: Url,
         destination: &Path,
-        file_size: Option<u64>,
         cancel: CancellationToken,
     ) -> Result<()> {
+        info!(
+            "downloading `{source}` to `{destination}`",
+            source = source.display(),
+            destination = destination.display(),
+        );
+
+        // Start by sending a HEAD request for the file's size and etag
+        let response = Retry::spawn_notify(
+            self.backend.config().retry_durations(),
+            || async {
+                select! {
+                    biased;
+                    _ = cancel.cancelled() => Err(Error::Canceled),
+                    r = self.backend.head(source.clone()) => r
+                }
+                .map_err(Error::into_retry_error)
+            },
+            notify_retry,
+        )
+        .await?;
+
+        let id = self.next_id();
+
+        // Check for a strong etag
+        let etag = response
+            .headers()
+            .get(header::ETAG)
+            .and_then(|v| v.to_str().ok())
+            .and_then(|v| {
+                if v.trim_start().starts_with("W/") {
+                    None
+                } else {
+                    Some(v)
+                }
+            });
+
+        // Get the content length
+        let content_length = response
+            .headers()
+            .get(header::CONTENT_LENGTH)
+            .and_then(|v| v.to_str().ok()?.parse().ok());
+
         let transfer = async {
             // Start by creating the destination's parent directory
             let parent = destination.parent().ok_or(Error::InvalidPath)?;
@@ -127,24 +133,16 @@ where
             let temp = NamedTempFile::with_prefix_in(".", parent)
                 .map_err(|error| Error::CreateTempFile { error })?;
 
-            let response = self.backend.get(source).await?;
-            let mut reader = StreamReader::new(TransferStream::new(
-                response.bytes_stream().map_err(std::io::Error::other),
+            // Download the file with resumable retries
+            self.download_with_resume(
                 id,
-                1,
-                self.backend.events().clone(),
-            ));
-
-            let mut writer = BufWriter::new(
-                fs::File::create(temp.path())
-                    .await
-                    .map_err(|error| Error::CreateTempFile { error })?,
-            );
-
-            // Copy the response stream to the temp file
-            tokio::io::copy(&mut reader, &mut writer)
-                .await
-                .map_err(Error::from)?;
+                &source,
+                temp.path(),
+                content_length,
+                etag,
+                cancel.clone(),
+            )
+            .await?;
 
             // Persist the temp file to the destination
             temp.persist(destination)
@@ -153,13 +151,14 @@ where
             Ok(())
         };
 
-        // Transfer as a single block
+        // Send the transfer started event
         if let Some(events) = self.backend.events() {
             events
-                .send(TransferEvent::BlockStarted {
+                .send(TransferEvent::TransferStarted {
                     id,
-                    block: 1,
-                    size: file_size,
+                    path: destination.to_path_buf(),
+                    blocks: 1,
+                    size: content_length,
                 })
                 .ok();
         }
@@ -170,11 +169,11 @@ where
             r = transfer => r,
         };
 
+        // Send the transfer end event
         if let Some(events) = self.backend.events() {
             events
-                .send(TransferEvent::BlockCompleted {
+                .send(TransferEvent::TransferCompleted {
                     id,
-                    block: 1,
                     failed: result.is_err(),
                 })
                 .ok();
@@ -183,54 +182,116 @@ where
         result
     }
 
-    /// Downloads a block of a file using a ranged GET request.
-    async fn download_block(
+    /// Downloads a file with resuming retry.
+    async fn download_with_resume(
         &self,
-        source: Url,
-        mut info: BlockDownloadInfo,
+        id: u64,
+        source: &Url,
+        destination: &Path,
+        content_length: Option<u64>,
+        etag: Option<&str>,
         cancel: CancellationToken,
     ) -> Result<()> {
-        let id = info.id;
-        let block = info.block;
-        let start = info.range.start;
-        let block_size = info.range.end - start;
-
         let transfer = async {
-            let response = self
-                .backend
-                .get_range(source, info.etag.as_str(), info.range)
-                .await?;
+            let offset = AtomicU64::new(0);
 
-            let mut reader = StreamReader::new(TransferStream::new(
-                response.bytes_stream().map_err(std::io::Error::other),
-                info.id,
-                info.block,
-                self.backend.events().clone(),
-            ));
+            // Retry the download with resume
+            Retry::spawn_notify(
+                self.backend.config().retry_durations(),
+                || async {
+                    let mut current = offset.load(Ordering::SeqCst);
 
-            // Read the block into the provided buffer
-            reader
-                .read_exact(&mut info.buffer[0..block_size as usize])
-                .map_err(Error::from)
-                .await?;
+                    let response = if current > 0
+                        && let Some(etag) = etag
+                    {
+                        self.backend
+                            .get_at_offset(source.clone(), etag, current)
+                            .await?
+                    } else {
+                        self.backend.get(source.clone()).await?
+                    };
 
-            // Write the block to the file
-            spawn_blocking(move || -> Result<_> {
-                crate::os::write_at(&info.file, &info.buffer[0..block_size as usize], start)
-                    .map_err(Error::from)
-            })
+                    // If the response is not partial, start from the beginning
+                    if current != 0
+                        && (etag.is_none() || response.status() != StatusCode::PARTIAL_CONTENT)
+                    {
+                        debug!(
+                            "resuming download of `{source}` from the beginning of the file",
+                            source = source.display()
+                        );
+                        current = 0;
+                    } else if response.status() == StatusCode::PARTIAL_CONTENT {
+                        // Ensure the response starts at the requested position
+                        if !response
+                            .headers()
+                            .get(header::CONTENT_RANGE)
+                            .and_then(|v| v.to_str().ok())
+                            .map(|v| v.trim_start().starts_with(&format!("bytes {current}-")))
+                            .unwrap_or(false)
+                        {
+                            return Err(Error::UnexpectedContentRangeStart.into());
+                        }
+
+                        debug!(
+                            "resuming download of `{source}` from offset {current}",
+                            source = source.display()
+                        );
+                    }
+
+                    let mut reader = StreamReader::new(TransferStream::new(
+                        response.bytes_stream().map_err(std::io::Error::other),
+                        id,
+                        0,
+                        current,
+                        self.backend.events().clone(),
+                    ));
+
+                    let mut file = fs::OpenOptions::new()
+                        .create(true)
+                        .write(true)
+                        .truncate(false)
+                        .open(destination)
+                        .await
+                        .map_err(|error| Error::CreateTempFile { error })?;
+
+                    file.set_len(current).await.map_err(Error::from)?;
+                    file.seek(SeekFrom::Start(current))
+                        .await
+                        .map_err(Error::from)?;
+
+                    let mut writer = BufWriter::new(file);
+
+                    // Copy the response stream to the temp file
+                    // If there is an error, update the current offset so that we resume on next
+                    // retry
+                    if let Err(e) = tokio::io::copy(&mut reader, &mut writer)
+                        .await
+                        .map_err(Error::from)
+                    {
+                        // Flush the writer and determine how much was written
+                        writer.flush().await.map_err(Error::from)?;
+                        let written = writer
+                            .seek(SeekFrom::Current(0))
+                            .await
+                            .map_err(Error::from)?;
+                        offset.store(written, Ordering::SeqCst);
+                        return Err(e.into());
+                    }
+
+                    Ok(())
+                },
+                notify_retry,
+            )
             .await
-            .expect("failed to join blocking task")?;
-
-            Ok(())
         };
 
+        // Send the block started event
         if let Some(events) = self.backend.events() {
             events
                 .send(TransferEvent::BlockStarted {
                     id,
-                    block,
-                    size: Some(block_size),
+                    block: 0,
+                    size: content_length,
                 })
                 .ok();
         }
@@ -241,11 +302,12 @@ where
             r = transfer => r,
         };
 
+        // Send the transfer end event
         if let Some(events) = self.backend.events() {
             events
                 .send(TransferEvent::BlockCompleted {
                     id,
-                    block,
+                    block: 0,
                     failed: result.is_err(),
                 })
                 .ok();
@@ -455,261 +517,45 @@ where
         )
         .await?;
 
-        // If there are no files relative to the given URL, download the URL as a file
+        // If there are no files relative to the given URL, download the URL a single
+        // file
         if paths.is_empty() {
-            return self.download_file(source, destination.to_path_buf()).await;
+            return self
+                .inner
+                .download(source, destination, self.cancel.clone())
+                .await;
         }
 
         // Otherwise, download each file in turn
-        for path in paths {
-            let mut source = source.clone();
-            let mut destination = destination.to_path_buf();
+        let mut downloads = stream::iter(paths)
+            .map(move |path| {
+                let mut source = source.clone();
+                let mut destination = destination.to_path_buf();
 
-            // Adjust source and destination based on the provided relative URL path
-            {
-                let mut segments = source.path_segments_mut().expect("URL should have a path");
-                for segment in path.split('/') {
-                    segments.push(segment);
-                    destination.push(segment);
-                }
-            }
-
-            self.download_file(source, destination).await?;
-        }
-
-        Ok(())
-    }
-
-    /// Downloads a file from the given URL.
-    async fn download_file(&self, source: Url, destination: PathBuf) -> Result<()> {
-        info!(
-            "downloading `{source}` to `{destination}`",
-            source = source.display(),
-            destination = destination.display(),
-        );
-
-        // Start by sending a HEAD request for the file
-        let response = Retry::spawn_notify(
-            self.inner.backend.config().retry_durations(),
-            || async {
-                select! {
-                    biased;
-                    _ = self.cancel.cancelled() => Err(Error::Canceled),
-                    r = self.inner.backend.head(source.clone()) => r
-                }
-                .map_err(Error::into_retry_error)
-            },
-            notify_retry,
-        )
-        .await?;
-
-        let file_size = response
-            .headers()
-            .get(header::CONTENT_LENGTH)
-            .and_then(|v| v.to_str().ok()?.parse().ok());
-
-        let accept_ranges = response
-            .headers()
-            .get(header::ACCEPT_RANGES)
-            .and_then(|v| v.to_str().ok());
-
-        let etag = response
-            .headers()
-            .get(header::ETAG)
-            .and_then(|v| v.to_str().ok());
-
-        let id = self.inner.next_id();
-
-        // Check to see if we can download the file with blocks
-        // For block downloads, the server must have responded with the resource size,
-        // accept byte ranges, and provided a strong etag (does not start with `W/`).
-        let result = match (file_size, accept_ranges, etag) {
-            (Some(file_size), Some("bytes"), Some(etag)) if !etag.starts_with("W/") => {
-                // Calculate the block size and the number of blocks to download
-                let block_size = self.inner.backend.block_size(file_size)?;
-                let num_blocks = file_size.div_ceil(block_size);
-
-                debug!(
-                    "file `{source}` is {file_size} bytes and will be downloaded with \
-                     {num_blocks} block(s) of size {block_size}",
-                    source = source.display()
-                );
-
-                if let Some(events) = self.inner.backend.events() {
-                    events
-                        .send(TransferEvent::TransferStarted {
-                            id,
-                            path: destination.clone(),
-                            blocks: num_blocks,
-                            size: Some(file_size),
-                        })
-                        .ok();
+                // Adjust source and destination based on the provided relative URL path
+                {
+                    let mut segments = source.path_segments_mut().expect("URL should have a path");
+                    for segment in path.split('/') {
+                        segments.push(segment);
+                        destination.push(segment);
+                    }
                 }
 
-                let info = DownloadInfo {
-                    id,
-                    file_size,
-                    block_size,
-                    num_blocks,
-                    etag: Arc::new(etag.to_string()),
-                };
-
-                // Download the file with multiple blocks
-                Retry::spawn_notify(
-                    self.inner.backend.config().retry_durations(),
-                    || {
-                        self.download_in_blocks(source.clone(), &destination, info.clone())
-                            .map_err(|e| match e {
-                                // Only retry on modified content errors; otherwise, we've already
-                                // internally retried downloads of the individual blocks
-                                Error::RemoteContentModified => RetryError::transient(e),
-                                _ => RetryError::permanent(e),
-                            })
-                    },
-                    notify_retry,
-                )
-                .await
-            }
-            _ => {
-                if file_size.is_none() {
-                    debug!(
-                        "file `{source}` will be downloaded as a single request because its size \
-                         could not be determined",
-                        source = source.display()
-                    );
-                } else if accept_ranges.is_none() {
-                    debug!(
-                        "file `{source}` will be downloaded as a single request because the \
-                         server does not accept ranged requests",
-                        source = source.display()
-                    );
-                } else if etag.is_none() {
-                    debug!(
-                        "file `{source}` will be downloaded as a single request because the \
-                         server did not respond with an ETAG",
-                        source = source.display()
-                    );
-                }
-
-                // Download the file as a single block
-                if let Some(events) = self.inner.backend.events() {
-                    events
-                        .send(TransferEvent::TransferStarted {
-                            id,
-                            path: destination.clone(),
-                            blocks: 1,
-                            size: file_size,
-                        })
-                        .ok();
-                }
-
-                Retry::spawn_notify(
-                    self.inner.backend.config().retry_durations(),
-                    || {
-                        self.inner
-                            .download(
-                                id,
-                                source.clone(),
-                                &destination,
-                                file_size,
-                                self.cancel.clone(),
-                            )
-                            .map_err(Error::into_retry_error)
-                    },
-                    notify_retry,
-                )
-                .await
-            }
-        };
-
-        if let Some(events) = self.inner.backend.events() {
-            events
-                .send(TransferEvent::TransferCompleted {
-                    id,
-                    failed: result.is_err(),
-                })
-                .ok();
-        }
-
-        result
-    }
-
-    /// Downloads a file in blocks using ranged GET requests.
-    ///
-    /// The blocks are downloaded in parallel and retried individually upon
-    /// errors, leading to faster error recovery.
-    async fn download_in_blocks(
-        &self,
-        source: Url,
-        destination: &Path,
-        info: DownloadInfo,
-    ) -> Result<()> {
-        // Start by creating the destination's parent directory
-        let parent = destination.parent().ok_or(Error::InvalidPath)?;
-        fs::create_dir_all(parent)
-            .await
-            .map_err(|error| Error::DirectoryCreationFailed {
-                path: parent.to_path_buf(),
-                error,
-            })?;
-
-        // Use a temp file that will be atomically renamed when the download completes
-        // Set its initial size up front so that individual writes don't go past EOF
-        let mut temp = NamedTempFile::with_prefix_in(".", parent)
-            .map_err(|error| Error::CreateTempFile { error })?;
-        temp.as_file_mut().set_len(info.file_size)?;
-
-        let (file, temp) = temp.into_parts();
-        let file = Arc::new(file);
-        let mut stream = stream::iter(0..info.num_blocks)
-            .map(|block| {
                 let inner = self.inner.clone();
-                let file = file.clone();
-                let source = source.clone();
-                let info = info.clone();
                 let cancel = self.cancel.clone();
-
-                tokio::spawn(async move {
-                    Retry::spawn_notify(
-                        inner.backend.config().retry_durations(),
-                        || {
-                            inner
-                                .download_block(
-                                    source.clone(),
-                                    BlockDownloadInfo {
-                                        id: info.id,
-                                        etag: info.etag.clone(),
-                                        file: file.clone(),
-                                        block,
-                                        buffer: inner.pool.alloc(info.block_size as usize),
-                                        range: (block * info.block_size)
-                                            ..((block + 1) * info.block_size).min(info.file_size),
-                                    },
-                                    cancel.clone(),
-                                )
-                                .map_err(Error::into_retry_error)
-                        },
-                        notify_retry,
-                    )
-                    .await
-                })
-                .map(|r| r.expect("task panicked"))
+                tokio::spawn(async move { inner.download(source, &destination, cancel).await })
+                    .map(|r| r.expect("task panicked"))
             })
             .buffer_unordered(self.inner.backend.config().parallelism());
 
         loop {
-            let result = stream.next().await;
+            let result = downloads.next().await;
             match result {
                 Some(r) => r?,
                 None => break,
             }
         }
 
-        drop(stream);
-
-        // Persist the temp file to the destination
-        temp.persist(destination)
-            .map_err(|e| Error::PersistTempFile { error: e.error })?;
         Ok(())
     }